feat(core): rename httpApi to apiGateway
