+++ packages/serverless-contracts/contracts/apiGateway/types.ts
