+++ packages/serverless-contracts/contracts/index.ts
@@ -1,2 +1,2 @@
-export { ApiGatewayContract } from './httpApi/apiGatewayContract';
+export { ApiGatewayContract } from './apiGateway/apiGatewayContract';
 export { CloudFormationContract } from './cloudFormation/cloudFormationContract';
